+++ lib/components/interfaces/touch_ripple_event.dart
@@ -1,0 +1,11 @@
+
+
+
+
+
+abstract class TouchRippleEvent {
+  
+  createCallback() {
+
+  }
+}