+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 ## 0.0.1
-* First varsion
+* First varsion (Inefficient)
 
 ## 1.0.0
-* New varsion
+* Rewritten the all code
+* modularised the code and relied on abstract classes.
+* made it extensible by implementing the gesture detector used in the package.
+* documented the code, and used efficient design patterns.
 
 ## 1.0.1
 * Bug fixes varsion
 
 ## 1.0.2
-* Updated package varsion
+* Updated package varsion (Optimised)
+
+## 1.0.3
+* Slightly updated version (For additional behaviors and gestures)
+
+<!--
+## 1.1.0
+Added the following events:
+* onVerticalDragStart, onVerticalDragUpdate onVerticalDragEnd, onHorizontalDragStart, onHorizontalDragUpdate, onHorizontalDragEnd.
+* Added new widgets: TouchRippleGestureDelegator, PrimaryTouchRippleController
+-->